Change account_age_in_months type to integer

+++ sde_dbt_tutorial/models/staging/stg_eltool__customers.sql
@@ -9,7 +9,7 @@ with source as (
         datetime_created::TIMESTAMP AS datetime_created,
         datetime_updated::TIMESTAMP AS datetime_updated,
         (DATE_PART('year', now()) - DATE_PART('year', datetime_created::TIMESTAMP)) * 12 +
-              (DATE_PART('month', now()) - DATE_PART('month', datetime_created::TIMESTAMP)) as account_age_in_months,
+              (DATE_PART('month', now()) - DATE_PART('month', datetime_created::TIMESTAMP))::integer as account_age_in_months,
         dbt_valid_from,
         dbt_valid_to
     from source
